Transitioned to DUT-centric namespace with aliases and bumped to v0.10.
John Parejko was having some trouble getting his thesis to work with
the previous implementation.  Apparently the standard approach to
theses for the astro folks at Drexel is to have most body chapters be
full-fledged papers from previous publications.  This lead to a bunch
of name collisions.  For example, the per-chapter abstracts were using
the drexel-thesis abstract environment, which was not appropriate.

I think the cut-and-paste-paper-chapter idea isn't all that great ;),
so I didn't want to just rename the drexel-thesis stuff to more
obscure names to avoid collisions.

The solution implemented here compromises, by adding DUT* aliases to
public macros and environments.  That way, most users can continue as
they were with no problems, but users who wish a macro or environment
name was different can clobber away and then use the DUT* form to
access the clobbered functionality.

I also moved some of the internal commands to drexel-thesis-specific
names, when I felt the internal commands were also a clobber risk.
The problem with this is that it becomes more awkward to override
the drexel-thesis behaviour, since then you *want* to successfully
clobber something.  If that ends up being a problem for someone,
let me know :).